+++ doomsday/plugins/jheretic/src/p_oldsvg.c
@@ -240,7 +240,7 @@ void P_v13_UnArchivePlayers(void)
 
 void P_v13_UnArchiveWorld(void)
 {
-    int         i, j;
+    uint        i, j;
     fixed_t     offx, offy;
     short      *get;
     int         firstflat = W_CheckNumForName("F_START") + 1;
@@ -506,7 +506,7 @@ void P_v13_UnArchiveSpecials(void)
  */
 void SV_v13_LoadGame(char *savename)
 {
-    int     length;
+    size_t  length;
     int     i;
     int     a, b, c;
     char    vcheck[VERSIONSIZE];
