+++ doomsday/plugins/jhexen/src/sc_man.c
@@ -86,8 +86,8 @@ static char *ScriptPtr;
 static char *ScriptEndPtr;
 static char StringBuffer[MAX_STRING_SIZE];
 static boolean ScriptOpen = false;
-static boolean ScriptFreeCLib;	// true = de-allocate using free()
-static int ScriptSize;
+static boolean ScriptFreeCLib;  // true = de-allocate using free()
+static size_t ScriptSize;
 static boolean AlreadyGot = false;
 
 // CODE --------------------------------------------------------------------
@@ -100,17 +100,17 @@ static boolean AlreadyGot = false;
 
 void SC_Open(char *name)
 {
-	char    fileName[128];
+    char    fileName[128];
 
-	if(sc_FileScripts == true)
-	{
-		sprintf(fileName, "%s%s.txt", sc_ScriptsDir, name);
-		SC_OpenFile(fileName);
-	}
-	else
-	{
-		SC_OpenLump(name);
-	}
+    if(sc_FileScripts == true)
+    {
+        sprintf(fileName, "%s%s.txt", sc_ScriptsDir, name);
+        SC_OpenFile(fileName);
+    }
+    else
+    {
+        SC_OpenLump(name);
+    }
 }
 
 //==========================================================================
@@ -123,7 +123,7 @@ void SC_Open(char *name)
 
 void SC_OpenLump(char *name)
 {
-	OpenScript(name, LUMP_SCRIPT);
+    OpenScript(name, LUMP_SCRIPT);
 }
 
 //==========================================================================
@@ -137,7 +137,7 @@ void SC_OpenLump(char *name)
 
 void SC_OpenFile(char *name)
 {
-	OpenScript(name, FILE_ZONE_SCRIPT);
+    OpenScript(name, FILE_ZONE_SCRIPT);
 }
 
 //==========================================================================
@@ -151,7 +151,7 @@ void SC_OpenFile(char *name)
 
 void SC_OpenFileCLib(char *name)
 {
-	OpenScript(name, FILE_CLIB_SCRIPT);
+    OpenScript(name, FILE_CLIB_SCRIPT);
 }
 
 //==========================================================================
@@ -162,33 +162,33 @@ void SC_OpenFileCLib(char *name)
 
 static void OpenScript(char *name, int type)
 {
-	SC_Close();
-	if(type == LUMP_SCRIPT)
-	{							// Lump script
-		ScriptBuffer = (char *) W_CacheLumpName(name, PU_STATIC);
-		ScriptSize = W_LumpLength(W_GetNumForName(name));
-		strcpy(ScriptName, name);
-		ScriptFreeCLib = false;	// De-allocate using gi.Z_Free()
-	}
-	else if(type == FILE_ZONE_SCRIPT)
-	{							// File script - zone
-		ScriptSize = M_ReadFile(name, (byte **) &ScriptBuffer);
-		M_ExtractFileBase(name, ScriptName);
-		ScriptFreeCLib = false;	// De-allocate using gi.Z_Free()
-	}
-	else
-	{							// File script - clib
-		ScriptSize = M_ReadFileCLib(name, (byte **) &ScriptBuffer);
-		M_ExtractFileBase(name, ScriptName);
-		ScriptFreeCLib = true;	// De-allocate using free()
-	}
-	ScriptPtr = ScriptBuffer;
-	ScriptEndPtr = ScriptPtr + ScriptSize;
-	sc_Line = 1;
-	sc_End = false;
-	ScriptOpen = true;
-	sc_String = StringBuffer;
-	AlreadyGot = false;
+    SC_Close();
+    if(type == LUMP_SCRIPT)
+    {                           // Lump script
+        ScriptBuffer = (char *) W_CacheLumpName(name, PU_STATIC);
+        ScriptSize = W_LumpLength(W_GetNumForName(name));
+        strcpy(ScriptName, name);
+        ScriptFreeCLib = false; // De-allocate using gi.Z_Free()
+    }
+    else if(type == FILE_ZONE_SCRIPT)
+    {                           // File script - zone
+        ScriptSize = M_ReadFile(name, (byte **) &ScriptBuffer);
+        M_ExtractFileBase(name, ScriptName);
+        ScriptFreeCLib = false; // De-allocate using gi.Z_Free()
+    }
+    else
+    {                           // File script - clib
+        ScriptSize = M_ReadFileCLib(name, (byte **) &ScriptBuffer);
+        M_ExtractFileBase(name, ScriptName);
+        ScriptFreeCLib = true;  // De-allocate using free()
+    }
+    ScriptPtr = ScriptBuffer;
+    ScriptEndPtr = ScriptPtr + ScriptSize;
+    sc_Line = 1;
+    sc_End = false;
+    ScriptOpen = true;
+    sc_String = StringBuffer;
+    AlreadyGot = false;
 }
 
 //==========================================================================
@@ -199,18 +199,18 @@ static void OpenScript(char *name, int type)
 
 void SC_Close(void)
 {
-	if(ScriptOpen)
-	{
-		if(ScriptFreeCLib == true)
-		{
-			free(ScriptBuffer);
-		}
-		else
-		{
-			Z_Free(ScriptBuffer);
-		}
-		ScriptOpen = false;
-	}
+    if(ScriptOpen)
+    {
+        if(ScriptFreeCLib == true)
+        {
+            free(ScriptBuffer);
+        }
+        else
+        {
+            Z_Free(ScriptBuffer);
+        }
+        ScriptOpen = false;
+    }
 }
 
 //==========================================================================
@@ -221,90 +221,90 @@ void SC_Close(void)
 
 boolean SC_GetString(void)
 {
-	char   *text;
-	boolean foundToken;
+    char   *text;
+    boolean foundToken;
 
-	CheckOpen();
-	if(AlreadyGot)
-	{
-		AlreadyGot = false;
-		return true;
-	}
-	foundToken = false;
-	sc_Crossed = false;
-	if(ScriptPtr >= ScriptEndPtr)
-	{
-		sc_End = true;
-		return false;
-	}
-	while(foundToken == false)
-	{
-		while(*ScriptPtr <= 32)
-		{
-			if(ScriptPtr >= ScriptEndPtr)
-			{
-				sc_End = true;
-				return false;
-			}
-			if(*ScriptPtr++ == '\n')
-			{
-				sc_Line++;
-				sc_Crossed = true;
-			}
-		}
-		if(ScriptPtr >= ScriptEndPtr)
-		{
-			sc_End = true;
-			return false;
-		}
-		if(*ScriptPtr != ASCII_COMMENT)
-		{						// Found a token
-			foundToken = true;
-		}
-		else
-		{						// Skip comment
-			while(*ScriptPtr++ != '\n')
-			{
-				if(ScriptPtr >= ScriptEndPtr)
-				{
-					sc_End = true;
-					return false;
+    CheckOpen();
+    if(AlreadyGot)
+    {
+        AlreadyGot = false;
+        return true;
+    }
+    foundToken = false;
+    sc_Crossed = false;
+    if(ScriptPtr >= ScriptEndPtr)
+    {
+        sc_End = true;
+        return false;
+    }
+    while(foundToken == false)
+    {
+        while(*ScriptPtr <= 32)
+        {
+            if(ScriptPtr >= ScriptEndPtr)
+            {
+                sc_End = true;
+                return false;
+            }
+            if(*ScriptPtr++ == '\n')
+            {
+                sc_Line++;
+                sc_Crossed = true;
+            }
+        }
+        if(ScriptPtr >= ScriptEndPtr)
+        {
+            sc_End = true;
+            return false;
+        }
+        if(*ScriptPtr != ASCII_COMMENT)
+        {                       // Found a token
+            foundToken = true;
+        }
+        else
+        {                       // Skip comment
+            while(*ScriptPtr++ != '\n')
+            {
+                if(ScriptPtr >= ScriptEndPtr)
+                {
+                    sc_End = true;
+                    return false;
 
-				}
-			}
-			sc_Line++;
-			sc_Crossed = true;
-		}
-	}
-	text = sc_String;
-	if(*ScriptPtr == ASCII_QUOTE)
-	{							// Quoted string
-		ScriptPtr++;
-		while(*ScriptPtr != ASCII_QUOTE)
-		{
-			*text++ = *ScriptPtr++;
-			if(ScriptPtr == ScriptEndPtr ||
-			   text == &sc_String[MAX_STRING_SIZE - 1])
-			{
-				break;
-			}
-		}
-		ScriptPtr++;
-	}
-	else
-	{							// Normal string
-		while((*ScriptPtr > 32) && (*ScriptPtr != ASCII_COMMENT))
-		{
-			*text++ = *ScriptPtr++;
-			if(ScriptPtr == ScriptEndPtr ||
-			   text == &sc_String[MAX_STRING_SIZE - 1])
-			{
-				break;
-			}
-		}
-	}
-	*text = 0;
-	return true;
+                }
+            }
+            sc_Line++;
+            sc_Crossed = true;
+        }
+    }
+    text = sc_String;
+    if(*ScriptPtr == ASCII_QUOTE)
+    {                           // Quoted string
+        ScriptPtr++;
+        while(*ScriptPtr != ASCII_QUOTE)
+        {
+            *text++ = *ScriptPtr++;
+            if(ScriptPtr == ScriptEndPtr ||
+               text == &sc_String[MAX_STRING_SIZE - 1])
+            {
+                break;
+            }
+        }
+        ScriptPtr++;
+    }
+    else
+    {                           // Normal string
+        while((*ScriptPtr > 32) && (*ScriptPtr != ASCII_COMMENT))
+        {
+            *text++ = *ScriptPtr++;
+            if(ScriptPtr == ScriptEndPtr ||
+               text == &sc_String[MAX_STRING_SIZE - 1])
+            {
+                break;
+            }
+        }
+    }
+    *text = 0;
+    return true;
 }
 
 //==========================================================================
@@ -315,10 +315,10 @@ boolean SC_GetString(void)
 
 void SC_MustGetString(void)
 {
-	if(SC_GetString() == false)
-	{
-		SC_ScriptError("Missing string.");
-	}
+    if(SC_GetString() == false)
+    {
+        SC_ScriptError("Missing string.");
+    }
 }
 
 //==========================================================================
@@ -329,11 +329,11 @@ void SC_MustGetString(void)
 
 void SC_MustGetStringName(char *name)
 {
-	SC_MustGetString();
-	if(SC_Compare(name) == false)
-	{
-		SC_ScriptError(NULL);
-	}
+    SC_MustGetString();
+    if(SC_Compare(name) == false)
+    {
+        SC_ScriptError(NULL);
+    }
 }
 
 //==========================================================================
@@ -344,23 +344,23 @@ void SC_MustGetStringName(char *name)
 
 boolean SC_GetNumber(void)
 {
-	char   *stopper;
+    char   *stopper;
 
-	CheckOpen();
-	if(SC_GetString())
-	{
-		sc_Number = strtol(sc_String, &stopper, 0);
-		if(*stopper != 0)
-		{
-			Con_Error("SC_GetNumber: Bad numeric constant \"%s\".\n"
-					  "Script %s, Line %d", sc_String, ScriptName, sc_Line);
-		}
-		return true;
-	}
-	else
-	{
-		return false;
-	}
+    CheckOpen();
+    if(SC_GetString())
+    {
+        sc_Number = strtol(sc_String, &stopper, 0);
+        if(*stopper != 0)
+        {
+            Con_Error("SC_GetNumber: Bad numeric constant \"%s\".\n"
+                      "Script %s, Line %d", sc_String, ScriptName, sc_Line);
+        }
+        return true;
+    }
+    else
+    {
+        return false;
+    }
 }
 
 //==========================================================================
@@ -371,10 +371,10 @@ boolean SC_GetNumber(void)
 
 void SC_MustGetNumber(void)
 {
-	if(SC_GetNumber() == false)
-	{
-		SC_ScriptError("Missing integer.");
-	}
+    if(SC_GetNumber() == false)
+    {
+        SC_ScriptError("Missing integer.");
+    }
 }
 
 //==========================================================================
@@ -387,7 +387,7 @@ void SC_MustGetNumber(void)
 
 void SC_UnGet(void)
 {
-	AlreadyGot = true;
+    AlreadyGot = true;
 }
 
 //==========================================================================
@@ -440,16 +440,16 @@ void SC_UnGet(void)
 
 int SC_MatchString(char **strings)
 {
-	int     i;
+    int     i;
 
-	for(i = 0; *strings != NULL; i++)
-	{
-		if(SC_Compare(*strings++))
-		{
-			return i;
-		}
-	}
-	return -1;
+    for(i = 0; *strings != NULL; i++)
+    {
+        if(SC_Compare(*strings++))
+        {
+            return i;
+        }
+    }
+    return -1;
 }
 
 //==========================================================================
@@ -460,14 +460,14 @@ int SC_MatchString(char **strings)
 
 int SC_MustMatchString(char **strings)
 {
-	int     i;
+    int     i;
 
-	i = SC_MatchString(strings);
-	if(i == -1)
-	{
-		SC_ScriptError(NULL);
-	}
-	return i;
+    i = SC_MatchString(strings);
+    if(i == -1)
+    {
+        SC_ScriptError(NULL);
+    }
+    return i;
 }
 
 //==========================================================================
@@ -478,11 +478,11 @@ int SC_MustMatchString(char **strings)
 
 boolean SC_Compare(char *text)
 {
-	if(strcasecmp(text, sc_String) == 0)
-	{
-		return true;
-	}
-	return false;
+    if(strcasecmp(text, sc_String) == 0)
+    {
+        return true;
+    }
+    return false;
 }
 
 //==========================================================================
@@ -493,12 +493,12 @@ boolean SC_Compare(char *text)
 
 void SC_ScriptError(char *message)
 {
-	if(message == NULL)
-	{
-		message = "Bad syntax.";
-	}
-	Con_Error("Script error, \"%s\" line %d: %s", ScriptName, sc_Line,
-			  message);
+    if(message == NULL)
+    {
+        message = "Bad syntax.";
+    }
+    Con_Error("Script error, \"%s\" line %d: %s", ScriptName, sc_Line,
+              message);
 }
 
 //==========================================================================
@@ -509,8 +509,8 @@ void SC_ScriptError(char *message)
 
 static void CheckOpen(void)
 {
-	if(ScriptOpen == false)
-	{
-		Con_Error("SC_ call before SC_Open().");
-	}
+    if(ScriptOpen == false)
+    {
+        Con_Error("SC_ call before SC_Open().");
+    }
 }
